Renamed tool directory.
