+++ src/integration/kotlin/kscript/integration/AnnotationTest.kt
@@ -1,9 +1,9 @@
 package kscript.integration
 
-import kscript.integration.tool.TestAssertion.any
-import kscript.integration.tool.TestAssertion.startsWith
-import kscript.integration.tool.TestAssertion.verify
-import kscript.integration.tool.TestContext.projectDir
+import kscript.integration.tools.TestAssertion.any
+import kscript.integration.tools.TestAssertion.startsWith
+import kscript.integration.tools.TestAssertion.verify
+import kscript.integration.tools.TestContext.projectDir
 import org.apache.commons.io.FileUtils
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
