+++ src/integration/kotlin/kscript/integration/EnvironmentTest.kt
@@ -1,8 +1,8 @@
 package kscript.integration
 
-import kscript.integration.tool.TestAssertion.startsWith
-import kscript.integration.tool.TestAssertion.verify
-import kscript.integration.tool.TestContext.projectDir
+import kscript.integration.tools.TestAssertion.startsWith
+import kscript.integration.tools.TestAssertion.verify
+import kscript.integration.tools.TestContext.projectDir
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
