+++ src/integration/kotlin/kscript/integration/IdeaTest.kt
@@ -1,9 +1,9 @@
 package kscript.integration
 
-import kscript.integration.tool.TestAssertion.any
-import kscript.integration.tool.TestAssertion.verify
-import kscript.integration.tool.TestContext.copyToExecutablePath
-import kscript.integration.tool.TestContext.projectDir
+import kscript.integration.tools.TestAssertion.any
+import kscript.integration.tools.TestAssertion.verify
+import kscript.integration.tools.TestContext.copyToExecutablePath
+import kscript.integration.tools.TestContext.projectDir
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
