+++ src/integration/kotlin/kscript/integration/ScriptInputModesTest.kt
@@ -1,9 +1,9 @@
 package kscript.integration
 
-import kscript.integration.tool.TestAssertion.startsWith
-import kscript.integration.tool.TestAssertion.verify
-import kscript.integration.tool.TestContext.projectDir
-import kscript.integration.tool.TestContext.testDir
+import kscript.integration.tools.TestAssertion.startsWith
+import kscript.integration.tools.TestAssertion.verify
+import kscript.integration.tools.TestContext.projectDir
+import kscript.integration.tools.TestContext.testDir
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
