+++ src/integration/kotlin/kscript/integration/SimpleTest.kt
@@ -1,7 +1,7 @@
 package kscript.integration
 
-import kscript.integration.tool.TestAssertion.startsWith
-import kscript.integration.tool.TestAssertion.verify
+import kscript.integration.tools.TestAssertion.startsWith
+import kscript.integration.tools.TestAssertion.verify
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
